added january tests and separated test objects
