+++ .scalafmt.conf
@@ -3,6 +3,10 @@ maxColumn = 140
 lineEndings = unix
 importSelectors = singleLine
 
+project.excludeFilters = [
+   "app/utils/ConfirmationTestCasesUtil.scala"
+]
+
 project {
   git = true
 }
